Moved FormInput.tsx to its proper folder
